+++ www/run/index.html
@@ -32,7 +32,7 @@
       <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/codemirror/5.48.4/codemirror.min.css" /> 
       <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet" crossorigin="anonymous"> 
       
-      <script src="https://kit.fontawesome.com/1a0e18d9d9.js" crossorigin="anonymous"></script>
+      <link rel="stylesheet" href="/css/all.css" id="treejs_styles">
       <script src="tree.js"></script>
       <link rel="stylesheet" href="treejs.css" id="treejs_styles">
 
@@ -47,8 +47,8 @@
       <script src="https://cdnjs.cloudflare.com/ajax/libs/codemirror/5.63.3/mode/python/python.min.js" crossorigin="anonymous" referrerpolicy="no-referrer"></script>
 
       <!-- These filenames are replaced by the build system -->
-      <script src="tulipcc-20250202061823.mjs" type="module"></script>
-      <script type="text/javascript" src="amy-20250202061823.js"></script>
+      <script src="tulipcc-20250202063841.mjs" type="module"></script>
+      <script type="text/javascript" src="amy-20250202063841.js"></script>
       <script type="text/javascript" src="spss.js"></script>
   </head>
